Add a flag to fluxd for requiring AWS detection
At present fluxd will check initially for access to the AWS API
(actually the EC2 metadata service), and log the error it gets if it's
not accessible.

For people running outside AWS, this can look like something is
failing, and cause undue anxiety. On the other hand, if you _are_
running in AWS you probably expect it to just work, and you _do_ want
to know about any failure -- otherwise, you might not notice that,
say, automation has stopped upgrading images.

To cover that second case, and defang the first, the flag
`--registry-ecr-mandatory` makes it a fatal error if fluxd can't reach
the EC2 metadata service (which I'm using as an approximation to being
able to access the AWS API in general). By using the flag, you can
escalate a silent failure into a crash, which -- we can hope -- you
are more likely to notice.