Salt clone lookup name with mirror name
This ensures the operator will detect changes to the git source URL in
a `HelmRelease`. Before this change the lookup was done by release name
only and once the operator had a working clone it would never look at
the git source URL again, resulting in a halt of rollouts for that
`HelmRelease`.